pyproject.toml
[tool.poetry]
authors = ["Harvard Library Innovation Lab"]
description = "Perma-specific Scoop REST API."
license = "MIT"
name = "perma-scoop-api"
packages = [{ include = "scoop_rest_api" }]
readme = "README.md"
version = "0.0.9"

[tool.poetry.dependencies]
#
# Core
#
Flask = "^2.3.2"
bcrypt = "^4.0.1"
billiard = "^4.2.0"
celery = { extras = ["redis"], version = "^5.4.0" }
cryptography = "^44.0.1"
gunicorn = "^23.0.0"
peewee = "^3.16.2"
python = "^3.11"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
#
# URL validation
#
netaddr = "^0.9.0"
playwright = "^1.40.0"
validators = "^0.34.0"
psycopg2 = "^2.9.9"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
pytest = "^7.3.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 100
